+++ SensingKitLib/src/main/java/org/sensingkit/sensingkitlib/data/SKAudioLevelData.java
@@ -21,6 +21,8 @@
 
 package org.sensingkit.sensingkitlib.data;
 
+import org.sensingkit.sensingkitlib.SKSensorModuleType;
+
 import java.util.Locale;
 
 public class SKAudioLevelData extends SKAbstractData {
@@ -32,7 +34,7 @@ public class SKAudioLevelData extends SKAbstractData {
 
     public SKAudioLevelData(long timestamp, int level) {
 
-        super(timestamp);
+        super(SKSensorModuleType.AUDIO_LEVEL, timestamp);
 
         this.level = level;
     }