+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Added support for Bluetooth sensor module.
   - DataInterface method getDataInString() has been renamed to getDataInCSV()
   - SKDataInterface has been renamed to SKSensorData
+  - Added getSensorModuleType() to SKSensorData
   - Added SK prefix to all source files
 
 ### 0.1.0